resturcture django webapp directories
Why: I have restructured the project directories so that the interop
client directory no long contains the code that will send data to the
drone. the drone client directory will now be the only directory that
contains code that will open up comunication directly with the drone. It
will also contain a bridge class that will help the drone client
interface with the interop client.

How: N/A